Move caseworker integration tests into event folder. Move response json files into responses folder
