+++ src/integrationTest/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerAddNoteIT.java
@@ -1,8 +1,10 @@
-package uk.gov.hmcts.sptribs.caseworker;
+package uk.gov.hmcts.sptribs.caseworker.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import feign.FeignException;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,31 +15,43 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
-import uk.gov.hmcts.sptribs.caseworker.model.HearingSummary;
-import uk.gov.hmcts.sptribs.caseworker.model.Listing;
 import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 import uk.gov.hmcts.sptribs.common.config.WebMvcConfig;
 import uk.gov.hmcts.sptribs.testutil.IdamWireMock;
 
+import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneId;
+
+import static java.util.Objects.requireNonNull;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.when;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static uk.gov.hmcts.sptribs.caseworker.util.EventConstants.CASEWORKER_PANEL_COMPOSITION;
-import static uk.gov.hmcts.sptribs.testutil.TestConstants.ABOUT_TO_START_URL;
+import static uk.gov.hmcts.sptribs.testutil.ITEventConstants.CASEWORKER_ADD_NOTE;
+import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.ST_CIC_CASEWORKER;
+import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.stubForIdamDetails;
+import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.stubForIdamFailure;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.AUTHORIZATION;
+import static uk.gov.hmcts.sptribs.testutil.TestConstants.CASEWORKER_USER_ID;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.SERVICE_AUTHORIZATION;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.TEST_AUTHORIZATION_TOKEN;
 import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.callbackRequest;
 import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.caseData;
+import static uk.gov.hmcts.sptribs.testutil.TestResourceUtil.expectedResponse;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
 @ContextConfiguration(initializers = {IdamWireMock.PropertiesInitializer.class})
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class CaseworkerPanelCompositionIT {
+class CaseworkerAddNoteIT {
+
+    private static final String CASEWORKER_ADD_NOTE_RESPONSE =
+        "classpath:responses/caseworker-add-note-response.json";
 
     @Autowired
     private MockMvc mockMvc;
@@ -47,6 +61,9 @@ public class CaseworkerPanelCompositionIT {
 
     @MockBean
     private WebMvcConfig webMvcConfig;
+
+    @MockBean
+    private Clock clock;
 
     @BeforeAll
     static void setUp() {
@@ -58,37 +75,18 @@ public class CaseworkerPanelCompositionIT {
         IdamWireMock.stopAndReset();
     }
 
-    @Test
-    void shouldPopulatePanel1OnAboutToStart() throws Exception {
-        final CaseData caseData = caseData();
-        final HearingSummary hearingSummary = HearingSummary.builder().build();
-        final Listing listing = Listing.builder()
-            .summary(hearingSummary)
-            .build();
-        caseData.setListing(listing);
-
-        mockMvc.perform(post(ABOUT_TO_START_URL)
-            .contentType(APPLICATION_JSON)
-            .header(SERVICE_AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-            .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-            .content(objectMapper.writeValueAsString(
-                callbackRequest(
-                    caseData,
-                    CASEWORKER_PANEL_COMPOSITION)))
-            .accept(APPLICATION_JSON))
-            .andExpect(
-                status().isOk())
-            .andExpect(jsonPath("$.data.panel1").value("Tribunal Judge"));
+    @BeforeEach
+    void setClock() {
+        when(clock.instant()).thenReturn(Instant.parse("2021-06-18T12:00:00.000Z"));
+        when(clock.getZone()).thenReturn(ZoneId.systemDefault());
     }
 
     @Test
-    void shouldPopulatePanelCompositionOnAboutToSubmit() throws Exception {
+    void shouldSuccessfullyAddCaseNoteToCaseDataWhenThereAreNoExistingCaseNotes() throws Exception {
         final CaseData caseData = caseData();
-        final HearingSummary hearingSummary = HearingSummary.builder().build();
-        final Listing listing = Listing.builder()
-            .summary(hearingSummary)
-            .build();
-        caseData.setListing(listing);
+        caseData.setNote("This is a test note");
+
+        stubForIdamDetails(TEST_AUTHORIZATION_TOKEN, CASEWORKER_USER_ID, ST_CIC_CASEWORKER);
 
         mockMvc.perform(post(ABOUT_TO_SUBMIT_URL)
             .contentType(APPLICATION_JSON)
@@ -97,10 +95,40 @@ public class CaseworkerPanelCompositionIT {
             .content(objectMapper.writeValueAsString(
                 callbackRequest(
                     caseData,
-                    CASEWORKER_PANEL_COMPOSITION)))
+                    CASEWORKER_ADD_NOTE)))
             .accept(APPLICATION_JSON))
             .andExpect(
                 status().isOk())
-            .andExpect(jsonPath("$.data.panelComposition").value("Panel 1"));
+            .andExpect(
+                content().json(expectedResponse(CASEWORKER_ADD_NOTE_RESPONSE))
+            );
+    }
+
+    @Test
+    void shouldThrowUnauthorizedExceptionWhenAboutToSubmitIsInvokedAndIdamThrowsUnauthorized() throws Exception {
+        final CaseData caseData = caseData();
+        caseData.setNote("This is a test note");
+
+        stubForIdamFailure();
+
+        mockMvc.perform(post(ABOUT_TO_SUBMIT_URL)
+            .contentType(APPLICATION_JSON)
+            .header(SERVICE_AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
+            .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
+            .content(objectMapper.writeValueAsString(
+                callbackRequest(
+                    caseData,
+                    CASEWORKER_ADD_NOTE)))
+            .accept(APPLICATION_JSON))
+            .andExpect(
+                status().isUnauthorized()
+            )
+            .andExpect(
+                result -> assertThat(result.getResolvedException()).isExactlyInstanceOf(FeignException.Unauthorized.class)
+            )
+            .andExpect(
+                result -> assertThat(requireNonNull(result.getResolvedException()).getMessage())
+                    .contains("Invalid idam credentials")
+            );
     }
 }
