+++ src/integrationTest/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerCaseFlagIT.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.sptribs.caseworker;
+package uk.gov.hmcts.sptribs.caseworker.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterAll;
@@ -13,40 +13,30 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
-import uk.gov.hmcts.sptribs.caseworker.model.Listing;
-import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 import uk.gov.hmcts.sptribs.common.config.WebMvcConfig;
+import uk.gov.hmcts.sptribs.common.service.CcdSupplementaryDataService;
 import uk.gov.hmcts.sptribs.testutil.IdamWireMock;
 
-import static java.util.Collections.emptySet;
 import static net.javacrumbs.jsonunit.assertj.JsonAssertions.assertThatJson;
-import static net.javacrumbs.jsonunit.assertj.JsonAssertions.json;
-import static net.javacrumbs.jsonunit.core.Option.IGNORING_EXTRA_FIELDS;
+import static org.mockito.Mockito.verify;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.YES;
-import static uk.gov.hmcts.sptribs.caseworker.util.EventConstants.CASEWORKER_CLEAR_HEARING_OPTIONS;
-import static uk.gov.hmcts.sptribs.ciccase.model.HearingFormat.FACE_TO_FACE;
-import static uk.gov.hmcts.sptribs.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
+import static uk.gov.hmcts.sptribs.caseworker.util.EventConstants.CASEWORKER_CASE_FLAG;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.AUTHORIZATION;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.SERVICE_AUTHORIZATION;
+import static uk.gov.hmcts.sptribs.testutil.TestConstants.SUBMITTED_URL;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.TEST_AUTHORIZATION_TOKEN;
+import static uk.gov.hmcts.sptribs.testutil.TestConstants.TEST_CASE_ID;
 import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.callbackRequest;
 import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.caseData;
-import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.getMockedHearingVenueData;
-import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.getMockedRegionData;
-import static uk.gov.hmcts.sptribs.testutil.TestResourceUtil.expectedResponse;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
 @ContextConfiguration(initializers = {IdamWireMock.PropertiesInitializer.class})
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class CaseworkerClearHearingOptionsIT {
-
-    private static final String CASEWORKER_CLEAR_HEARING_OPTIONS_RESPONSE =
-        "classpath:caseworker-clear-hearing-options-response.json";
+public class CaseworkerCaseFlagIT {
 
     @Autowired
     private MockMvc mockMvc;
@@ -56,6 +46,11 @@ public class CaseworkerClearHearingOptionsIT {
 
     @MockBean
     private WebMvcConfig webMvcConfig;
+
+    @MockBean
+    private CcdSupplementaryDataService coreCaseApiService;
+
+    private static final String CONFIRMATION_HEADER = "$.confirmation_header";
 
     @BeforeAll
     static void setUp() {
@@ -68,27 +63,15 @@ public class CaseworkerClearHearingOptionsIT {
     }
 
     @Test
-    void shouldClearPreviouslyPopulatedHearingOptionsOnAboutToSubmit() throws Exception {
-        final CaseData caseData = caseData();
-        final Listing listing = Listing.builder()
-            .regionList(getMockedRegionData())
-            .hearingVenues(getMockedHearingVenueData())
-            .venueNotListedOption(emptySet())
-            .roomAtVenue("G.01")
-            .addlInstr("Ground floor")
-            .hearingFormat(FACE_TO_FACE)
-            .shortNotice(YES)
-            .build();
-        caseData.setListing(listing);
-
-        String response = mockMvc.perform(post(ABOUT_TO_SUBMIT_URL)
+    void shouldReturnErrorMessageIfNotificationsFailOnSubmitted() throws Exception {
+        String response = mockMvc.perform(post(SUBMITTED_URL)
             .contentType(APPLICATION_JSON)
             .header(SERVICE_AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
             .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
             .content(objectMapper.writeValueAsString(
                 callbackRequest(
-                    caseData,
-                    CASEWORKER_CLEAR_HEARING_OPTIONS)))
+                    caseData(),
+                    CASEWORKER_CASE_FLAG)))
             .accept(APPLICATION_JSON))
             .andExpect(
                 status().isOk())
@@ -97,7 +80,10 @@ public class CaseworkerClearHearingOptionsIT {
             .getContentAsString();
 
         assertThatJson(response)
-            .when(IGNORING_EXTRA_FIELDS)
-            .isEqualTo(json(expectedResponse(CASEWORKER_CLEAR_HEARING_OPTIONS_RESPONSE)));
+            .inPath(CONFIRMATION_HEADER)
+            .isString()
+            .contains("# Flag created \n## This Flag has been added to case");
+
+        verify(coreCaseApiService).submitSupplementaryDataToCcd(TEST_CASE_ID.toString());
     }
 }
