+++ src/integrationTest/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerCloseTheCaseIT.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.sptribs.caseworker;
+package uk.gov.hmcts.sptribs.caseworker.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterAll;
@@ -80,9 +80,9 @@ import static uk.gov.hmcts.sptribs.testutil.TestResourceUtil.expectedResponse;
 public class CaseworkerCloseTheCaseIT {
 
     private static final String CASEWORKER_CLOSE_THE_CASE_ABOUT_TO_START_RESPONSE =
-        "classpath:caseworker-close-the-case-about-to-start-response.json";
+        "classpath:responses/caseworker-close-the-case-about-to-start-response.json";
     private static final String CASEWORKER_CLOSE_THE_CASE_ABOUT_TO_SUBMIT_RESPONSE =
-        "classpath:caseworker-close-the-case-about-to-submit-response.json";
+        "classpath:responses/caseworker-close-the-case-about-to-submit-response.json";
     @Autowired
     private MockMvc mockMvc;
