+++ src/integrationTest/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerCreateBundleIT.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.sptribs.caseworker;
+package uk.gov.hmcts.sptribs.caseworker.event;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -58,7 +58,7 @@ import static uk.gov.hmcts.sptribs.testutil.TestResourceUtil.expectedResponse;
 public class CaseworkerCreateBundleIT {
 
     private static final String CASEWORKER_CREATE_BUNDLE_ABOUT_TO_SUBMIT_RESPONSE =
-        "classpath:caseworker-create-bundle-about-to-submit-response.json";
+        "classpath:responses/caseworker-create-bundle-about-to-submit-response.json";
 
     @Autowired
     private MockMvc mockMvc;
