+++ src/integrationTest/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerCreateHearingSummaryIT.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.sptribs.caseworker;
+package uk.gov.hmcts.sptribs.caseworker.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterAll;
@@ -71,9 +71,9 @@ import static uk.gov.hmcts.sptribs.testutil.TestResourceUtil.expectedResponse;
 public class CaseworkerCreateHearingSummaryIT {
 
     private static final String CASEWORKER_CREATE_HEARING_SUMMARY_ABOUT_TO_START_RESPONSE =
-        "classpath:caseworker-create-hearing-summary-about-to-start-response.json";
+        "classpath:responses/caseworker-create-hearing-summary-about-to-start-response.json";
     private static final String CASEWORKER_CREATE_HEARING_SUMMARY_ABOUT_TO_SUBMIT_RESPONSE =
-        "classpath:caseworker-create-hearing-summary-about-to-submit-response.json";
+        "classpath:responses/caseworker-create-hearing-summary-about-to-submit-response.json";
 
     @Autowired
     private MockMvc mockMvc;
