+++ src/integrationTest/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerIssueCaseIT.java
@@ -1,10 +1,8 @@
-package uk.gov.hmcts.sptribs.caseworker;
+package uk.gov.hmcts.sptribs.caseworker.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import feign.FeignException;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,25 +13,18 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
+import uk.gov.hmcts.sptribs.caseworker.model.CaseIssue;
 import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 import uk.gov.hmcts.sptribs.common.config.WebMvcConfig;
 import uk.gov.hmcts.sptribs.testutil.IdamWireMock;
 
-import java.time.Clock;
-import java.time.Instant;
-import java.time.ZoneId;
-
-import static java.util.Objects.requireNonNull;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.when;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static uk.gov.hmcts.sptribs.testutil.ITEventConstants.CASEWORKER_ADD_NOTE;
+import static uk.gov.hmcts.sptribs.testutil.ITEventConstants.CASEWORKER_ISSUE_CASE;
 import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.ST_CIC_CASEWORKER;
 import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.stubForIdamDetails;
-import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.stubForIdamFailure;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.AUTHORIZATION;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.CASEWORKER_USER_ID;
@@ -48,10 +39,10 @@ import static uk.gov.hmcts.sptribs.testutil.TestResourceUtil.expectedResponse;
 @AutoConfigureMockMvc
 @ContextConfiguration(initializers = {IdamWireMock.PropertiesInitializer.class})
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class CaseworkerAddNoteIT {
+class CaseworkerIssueCaseIT {
 
-    private static final String CASEWORKER_ADD_NOTE_RESPONSE =
-        "classpath:caseworker-add-note-response.json";
+    private static final String CASEWORKER_ISSUE_CASE_RESPONSE =
+        "classpath:responses/caseworker-issue-case-response.json";
 
     @Autowired
     private MockMvc mockMvc;
@@ -62,8 +53,6 @@ class CaseworkerAddNoteIT {
     @MockBean
     private WebMvcConfig webMvcConfig;
 
-    @MockBean
-    private Clock clock;
 
     @BeforeAll
     static void setUp() {
@@ -75,60 +64,28 @@ class CaseworkerAddNoteIT {
         IdamWireMock.stopAndReset();
     }
 
-    @BeforeEach
-    void setClock() {
-        when(clock.instant()).thenReturn(Instant.parse("2021-06-18T12:00:00.000Z"));
-        when(clock.getZone()).thenReturn(ZoneId.systemDefault());
-    }
 
     @Test
-    void shouldSuccessfullyAddCaseNoteToCaseDataWhenThereAreNoExistingCaseNotes() throws Exception {
+    void shouldSuccessfullyAIssueCase() throws Exception {
         final CaseData caseData = caseData();
-        caseData.setNote("This is a test note");
+        caseData.setCaseIssue(new CaseIssue());
 
         stubForIdamDetails(TEST_AUTHORIZATION_TOKEN, CASEWORKER_USER_ID, ST_CIC_CASEWORKER);
 
         mockMvc.perform(post(ABOUT_TO_SUBMIT_URL)
-            .contentType(APPLICATION_JSON)
-            .header(SERVICE_AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-            .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-            .content(objectMapper.writeValueAsString(
-                callbackRequest(
-                    caseData,
-                    CASEWORKER_ADD_NOTE)))
-            .accept(APPLICATION_JSON))
+                .contentType(APPLICATION_JSON)
+                .header(SERVICE_AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
+                .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
+                .content(objectMapper.writeValueAsString(
+                    callbackRequest(
+                        caseData,
+                        CASEWORKER_ISSUE_CASE)))
+                .accept(APPLICATION_JSON))
             .andExpect(
                 status().isOk())
             .andExpect(
-                content().json(expectedResponse(CASEWORKER_ADD_NOTE_RESPONSE))
+                content().json(expectedResponse(CASEWORKER_ISSUE_CASE_RESPONSE))
             );
     }
 
-    @Test
-    void shouldThrowUnauthorizedExceptionWhenAboutToSubmitIsInvokedAndIdamThrowsUnauthorized() throws Exception {
-        final CaseData caseData = caseData();
-        caseData.setNote("This is a test note");
-
-        stubForIdamFailure();
-
-        mockMvc.perform(post(ABOUT_TO_SUBMIT_URL)
-            .contentType(APPLICATION_JSON)
-            .header(SERVICE_AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-            .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-            .content(objectMapper.writeValueAsString(
-                callbackRequest(
-                    caseData,
-                    CASEWORKER_ADD_NOTE)))
-            .accept(APPLICATION_JSON))
-            .andExpect(
-                status().isUnauthorized()
-            )
-            .andExpect(
-                result -> assertThat(result.getResolvedException()).isExactlyInstanceOf(FeignException.Unauthorized.class)
-            )
-            .andExpect(
-                result -> assertThat(requireNonNull(result.getResolvedException()).getMessage())
-                    .contains("Invalid idam credentials")
-            );
-    }
 }
