+++ src/integrationTest/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerPanelCompositionIT.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.sptribs.caseworker;
+package uk.gov.hmcts.sptribs.caseworker.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterAll;
@@ -13,36 +13,31 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
-import uk.gov.hmcts.sptribs.caseworker.model.CaseIssue;
+import uk.gov.hmcts.sptribs.caseworker.model.HearingSummary;
+import uk.gov.hmcts.sptribs.caseworker.model.Listing;
 import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 import uk.gov.hmcts.sptribs.common.config.WebMvcConfig;
 import uk.gov.hmcts.sptribs.testutil.IdamWireMock;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static uk.gov.hmcts.sptribs.testutil.ITEventConstants.CASEWORKER_ISSUE_CASE;
-import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.ST_CIC_CASEWORKER;
-import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.stubForIdamDetails;
+import static uk.gov.hmcts.sptribs.caseworker.util.EventConstants.CASEWORKER_PANEL_COMPOSITION;
+import static uk.gov.hmcts.sptribs.testutil.TestConstants.ABOUT_TO_START_URL;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.AUTHORIZATION;
-import static uk.gov.hmcts.sptribs.testutil.TestConstants.CASEWORKER_USER_ID;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.SERVICE_AUTHORIZATION;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.TEST_AUTHORIZATION_TOKEN;
 import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.callbackRequest;
 import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.caseData;
-import static uk.gov.hmcts.sptribs.testutil.TestResourceUtil.expectedResponse;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
 @ContextConfiguration(initializers = {IdamWireMock.PropertiesInitializer.class})
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class CaseworkerIssueCaseIT {
-
-    private static final String CASEWORKER_ISSUE_CASE_RESPONSE =
-        "classpath:caseworker-issue-case-response.json";
+public class CaseworkerPanelCompositionIT {
 
     @Autowired
     private MockMvc mockMvc;
@@ -52,7 +47,6 @@ class CaseworkerIssueCaseIT {
 
     @MockBean
     private WebMvcConfig webMvcConfig;
-
 
     @BeforeAll
     static void setUp() {
@@ -64,28 +58,49 @@ class CaseworkerIssueCaseIT {
         IdamWireMock.stopAndReset();
     }
 
-
     @Test
-    void shouldSuccessfullyAIssueCase() throws Exception {
+    void shouldPopulatePanel1OnAboutToStart() throws Exception {
         final CaseData caseData = caseData();
-        caseData.setCaseIssue(new CaseIssue());
+        final HearingSummary hearingSummary = HearingSummary.builder().build();
+        final Listing listing = Listing.builder()
+            .summary(hearingSummary)
+            .build();
+        caseData.setListing(listing);
 
-        stubForIdamDetails(TEST_AUTHORIZATION_TOKEN, CASEWORKER_USER_ID, ST_CIC_CASEWORKER);
-
-        mockMvc.perform(post(ABOUT_TO_SUBMIT_URL)
-                .contentType(APPLICATION_JSON)
-                .header(SERVICE_AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-                .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
-                .content(objectMapper.writeValueAsString(
-                    callbackRequest(
-                        caseData,
-                        CASEWORKER_ISSUE_CASE)))
-                .accept(APPLICATION_JSON))
+        mockMvc.perform(post(ABOUT_TO_START_URL)
+            .contentType(APPLICATION_JSON)
+            .header(SERVICE_AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
+            .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
+            .content(objectMapper.writeValueAsString(
+                callbackRequest(
+                    caseData,
+                    CASEWORKER_PANEL_COMPOSITION)))
+            .accept(APPLICATION_JSON))
             .andExpect(
                 status().isOk())
-            .andExpect(
-                content().json(expectedResponse(CASEWORKER_ISSUE_CASE_RESPONSE))
-            );
+            .andExpect(jsonPath("$.data.panel1").value("Tribunal Judge"));
     }
 
+    @Test
+    void shouldPopulatePanelCompositionOnAboutToSubmit() throws Exception {
+        final CaseData caseData = caseData();
+        final HearingSummary hearingSummary = HearingSummary.builder().build();
+        final Listing listing = Listing.builder()
+            .summary(hearingSummary)
+            .build();
+        caseData.setListing(listing);
+
+        mockMvc.perform(post(ABOUT_TO_SUBMIT_URL)
+            .contentType(APPLICATION_JSON)
+            .header(SERVICE_AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
+            .header(AUTHORIZATION, TEST_AUTHORIZATION_TOKEN)
+            .content(objectMapper.writeValueAsString(
+                callbackRequest(
+                    caseData,
+                    CASEWORKER_PANEL_COMPOSITION)))
+            .accept(APPLICATION_JSON))
+            .andExpect(
+                status().isOk())
+            .andExpect(jsonPath("$.data.panelComposition").value("Panel 1"));
+    }
 }
