+++ src/integrationTest/java/uk/gov/hmcts/sptribs/citizen/event/CicDssUpdateCaseEventIT.java
@@ -56,7 +56,7 @@ import static uk.gov.hmcts.sptribs.testutil.TestResourceUtil.expectedResponse;
 public class CicDssUpdateCaseEventIT {
 
     private static final String CITIZEN_DSS_UPDATE_CASE_ABOUT_TO_SUBMIT_RESPONSE =
-        "classpath:citizen-dss-update-case-about-to-submit-response.json";
+        "classpath:responses/citizen-dss-update-case-about-to-submit-response.json";
 
     @Autowired
     private MockMvc mockMvc;
