+++ src/integrationTest/resources/responses/caseworker-cancel-hearing-about-to-start-response.json
