+++ src/integrationTest/resources/responses/caseworker-cancel-hearing-about-to-submit-response.json
