+++ src/integrationTest/resources/responses/caseworker-clear-hearing-options-response.json
