+++ src/integrationTest/resources/responses/caseworker-close-the-case-about-to-start-response.json
