+++ src/integrationTest/resources/responses/caseworker-close-the-case-about-to-submit-response.json
