+++ src/integrationTest/resources/responses/caseworker-create-draft-order-mid-event-response.json
