+++ src/integrationTest/resources/responses/caseworker-create-hearing-summary-about-to-submit-response.json
