+++ src/integrationTest/resources/responses/caseworker-hearing-options-mid-event-response.json
