+++ src/integrationTest/resources/responses/citizen-create-case-about-to-submit-response.json
