+++ src/integrationTest/resources/responses/citizen-create-case-response.json
