+++ src/integrationTest/resources/responses/citizen-document-upload-response.json
