+++ src/integrationTest/resources/responses/citizen-dss-update-case-about-to-submit-response.json
